stream: split text buffers in two: html and native
